pyproject.toml
[project]
name = "tesseract"
version = "0.1.0"
description = "Universal Atomic Composability for Rollups"
authors = [{name = "Dipankar Sarkar", email = "me@dipankar.name"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "vyper>=0.3.10,<0.4.0",
    "web3>=6.11.0,<6.12.0",
    "pytest>=7.0.0,<8.0.0",
    "python-dotenv>=1.0.0",
    "eth-utils>=2.0.0,<3.0.0",
    "eth-tester[py-evm]>=0.12.0b1,<0.13.0",
    "eth-typing>=3.0.0,<4.0.0",
    "setuptools>=80.9.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "pytest-cov>=4.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tesseract"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "security: Security tests",
    "slow: Slow-running tests"
]

[tool.coverage.run]
source = ["tesseract", "scripts"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError"
]
fail_under = 80
show_missing = true
